Again not providing __init__.py - bad precedent to set.
As much as I want to provide a broad and "Right(TM)" way to do it,
it is not clear who if anyone actually needs it, and that it will
vary too much by inidividual compatibility requirements. I don't
want to support any one ``__init__.py`` as part of this unless it
becomes obviously helpful enough to enough people to be worth it.